Add eyesight gameobject
Adds EyeSight game object to the player object, I think the idea was that it multiplies things in shadow by 0, making all colors black, which means things still render, such as someone behind a wall with a light. the light still updates and bleeds around the corner, but you can't see where its coming from directly.

This currently has a couple bugs, need to adjust layer order and lighting order for a lot of things